requirements.txt
astroid==1.5.3
certifi==2017.7.27.1
click==6.7
colorama==0.3.9
Flask==0.12.2
future==0.16.0
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.7.2
python-telegram-bot==7.0.1
six==1.10.0
virtualenv==15.1.0
Werkzeug==0.12.2
wrapt==1.10.10
